add minor compatibility messages (kbx), dedup dirs directories are now only referenced in one place (TokenDir in token/token.go, keyring locations in gpg/gpg.go), meaning it will be easier to in the future generalize/customize these.
